Explicitly specify busybox date command
bearDropper.sh currently fails on line 104 when coreutils-date is installed, because the GNU date command does not support the -D option.

The solution is to explicitly specify the busybox date command. I think the change is only actually needed on line 104, but I've changed all the date commands for consistency.